+++ CHANGES.md
@@ -12,6 +12,8 @@ libcups v3.0b1 (Month DD, YYYY)
 - Now require POSIX or Windows threading support.
 - Now require the `poll` function (`WSAPoll` on Windows).
 - Added new `GENERATE-FILE` directive for `ipptool` test files.
+- Added new `ATTR-IF-DEFINED` and `ATTR-IF-NOT-DEFINED` directives to IPP data
+  files (Issue #3)
 - Added `ippFile` API for working with IPP data files as used by `ipptool`,
   `ippexeprinter`, and other tools (Issue #14)
 - Updated the CUPS API for consistency.